Abstract alignable trait and create layout widget
