+++ src/ui/core/mod.rs
@@ -2,17 +2,20 @@
 // ongoing rust bug, check #18241 to see if they're resolved,
 // if so remove pub!
 pub mod painter;
-pub mod layout;
+pub mod alignable;
 pub mod widget;
 pub mod button;
 pub mod base;
+pub mod layout;
 
 // Because of the bug, there's no use showing these re-exports in the docs, 
 // so hide them all
 #[doc(hidden)]
 pub use ui::core::painter::Painter;
 #[doc(hidden)]
-pub use ui::core::layout::{Alignable, HorizontalAlign, VerticalAlign, HorizontalLayout};
+pub use ui::core::alignable::{Alignable, HorizontalAlign, VerticalAlign};
+#[doc(hidden)]
+pub use ui::core::layout::Layout;
 #[doc(hidden)]
 pub use ui::core::widget::Widget;
 #[doc(hidden)]
