+++ src/ui/mod.rs
@@ -1,6 +1,8 @@
 mod dialog;
 mod stdbutton;
+mod hlayout;
 pub mod core;
 
 pub use ui::dialog::Dialog;
 pub use ui::stdbutton::StdButton;
+pub use ui::hlayout::HorizontalLayout;
